Various fixes and tweaks to make it compiled under MinGW-w64 i686.
Patches retrieved from:

https://github.com/ralish/kitty/commit/901a529efa24a68e5e40cf0939cfc4e967840e59
https://github.com/Alexpux/MINGW-packages/commit/9ca49687d3861c44c610e8bb1c6bb0cd85528251

+++ 0.69_My_PuTTY/windows/winnet.c
@@ -145,7 +145,7 @@ static int curpass;
 static int cmpfortree(void *av, void *bv)
 {
     Actual_Socket a = (Actual_Socket) av, b = (Actual_Socket) bv;
-    unsigned long as = (unsigned long) a->s, bs = (unsigned long) b->s;
+    uintptr_t as = (uintptr_t) a->s, bs = (uintptr_t) b->s;
     if (as < bs)
 	return -1;
     if (as > bs)
